# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Brownie API
  version: "0.1.0"
  description: >
    The public API surface under /api/v1. It includes the platform
    self-check routes and the implemented workspace document, job, and
    persisted progress-event routes.

    Three answers are described here once instead of under each route,
    all in the Error shape. Any route can answer 429 RATE_LIMITED with a
    Retry-After header, when one person (or, for someone not signed in,
    one address) has made more requests of one kind in a minute than
    allowed, and 503 DATABASE_UNAVAILABLE with a Retry-After header. Every
    route that looks up the person asking, which is all of them but
    /api/v1/capabilities and /api/v1/data-practices, can answer 401
    SESSION_NO_LONGER_VALID, when the session is genuine but the person it
    names has no identity record any more, in which case the session has
    been ended. A request body larger than the API reads, other than a
    file upload, is answered 413 CONTENT_TOO_LARGE whatever kind of body
    it says it is.
servers:
  - url: http://localhost:8081
    description: Local development
paths:
...
  /api/v1/workspaces/{workspaceId}/documents/{documentId}:
    get:
      operationId: getDocument
      summary: Get a document and its selected immutable revision.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/workspaces/{workspaceId}/documents/{documentId}/export-approval:
    post:
      operationId: approveExport
      summary: >
        Approve an already-clean validation manifest for export. Refused
        if the manifest is stale against the document's current revision
        or still has an unresolved blocking finding. Approving again what
        is already the document's latest approval (same person, manifest
        and format) answers with that approval, so a repeated request
        records nothing twice. Any other approval, including going back to
        a format chosen earlier, is a new one and becomes the latest, which
        is the one an export follows. A PDF form is exported as a PDF only:
        approving DOCX or BOTH for one answers 422 with code
        EXPORT_FORMAT_NOT_OFFERED.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveExportRequest"
      responses:
        "201":
          description: Approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportApprovalResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
    get:
      operationId: getLatestExportApproval
      summary: The latest export approval recorded for this document, if any.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportApprovalResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/workspaces/{workspaceId}/documents/{documentId}/export:
    post:
      operationId: exportDocument
      summary: >
        Produce an export receipt from the document's latest export
        approval. Refused if that approval is stale against the current
        revision, or if a named artifact no longer matches the hash its
        own validation manifest recorded. Exporting the same approval again
        answers with the receipt that already exists, so a repeated request
        issues no second receipt and no second audit entry.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "201":
          description: Exported.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportReceiptResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/workspaces/{workspaceId}/documents/{documentId}/export-receipt:
    get:
      operationId: getLatestExportReceipt
      summary: The latest export receipt recorded for this document, if any.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportReceiptResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
...
    NotFound:
      description: The requested resource is not visible in this workspace.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
    UnprocessableContent:
      description: The typed request is well formed but semantically invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    PreconditionFailed:
      description: The named approval or manifest is stale against the document's current revision.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: >
        Something the service depends on is away, and the same request is
        worth making again shortly; nothing was lost or changed. The code
        says what: STORAGE_UNAVAILABLE (files cannot be stored or read),
        SCANNER_UNAVAILABLE (an upload cannot be scanned yet, and stays
        unreadable until it is), RENDERER_BUSY (every slot for preparing a
        document stayed taken; with a Retry-After header), or
        DATABASE_UNAVAILABLE, which any route at all can answer and which is
        therefore not repeated under each one.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    FieldValueResponse:
      type: object
      required: [type, cardinality]
      properties:
        type:
          type: string
          enum: [TEXT, DATE]
        cardinality:
          type: string
          enum: [SCALAR, REPEATED]
        value:
          type: string
          nullable: true
        values:
          type: array
          items:
            type: string
          nullable: true
        evidenceSourceSpanIds:
          type: array
          items:
            type: integer
            format: int64
        fieldState:
          allOf:
            - $ref: "#/components/schemas/FieldStateResponse"
          nullable: true
          description: Present for a SCALAR field.
        itemFieldStates:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/FieldStateResponse"
            nullable: true
          nullable: true
          description: Present for a REPEATED field, one entry per item in the same order as values.
    FieldStateResponse:
      type: object
      required: [authorship, evidenceSupport, validation, review, lock]
      properties:
        authorship:
          type: string
          enum: [IMPORTED, AI_COMPOSED, USER_AUTHORED, MIXED]
        evidenceSupport:
          type: string
          enum: [DIRECT, TRANSFORMED, AMBIGUOUS, UNSUPPORTED, MISSING]
        validation:
          type: string
          enum: [NOT_RUN, PASSED, WARNING, BLOCKING, UNAVAILABLE]
        review:
          type: string
          enum: [UNREVIEWED, ACCEPTED, REJECTED, NEEDS_CLARIFICATION]
        lock:
          type: string
          enum: [EDITABLE, PRESERVE_ON_REGENERATION, EXPLICITLY_LOCKED]
    DocumentRevisionResponse:
      type: object
      required: [id, revisionNumber, fields, contentHash, editReason, createdAt]
      properties:
        id:
          type: integer
          format: int64
        revisionNumber:
          type: integer
          minimum: 1
        parentRevisionId:
          type: integer
          format: int64
          nullable: true
        fields:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FieldValueResponse"
        contentHash:
          type: string
          pattern: "^[0-9a-f]{64}$"
        editReason:
          type: string
        createdAt:
          type: string
          format: date-time
        templateVersionId:
          type: integer
          format: int64
          description: >
            The template version this revision's content was written against;
            always sent by a server that can move documents between versions.
...
    DocumentResponse:
      type: object
      required: [id, title, templateId, templateVersionId, currentRevisionId, createdAt, currentRevision]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        templateId:
          type: integer
          format: int64
        templateVersionId:
          type: integer
          format: int64
        templateLatestVersionId:
          type: integer
          format: int64
          nullable: true
          description: >
            The template's current version, the one new documents start from.
            When it differs from templateVersionId the form has a newer
            version this document can be moved to.
        currentRevisionId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        currentRevision:
          $ref: "#/components/schemas/DocumentRevisionResponse"
...
    Error:
      description: >
        The shape of every error response in this API, regardless of
        cause.
      type: object
      required: [status, title, code, correlationId, fields, recoveryActions]
      properties:
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable, machine-readable error code (e.g. NOT_FOUND, VALIDATION_FAILED).
        correlationId:
          type: string
          description: Echoes the X-Correlation-Id request/response header.
        fields:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        recoveryActions:
          type: array
          items:
            type: string
        limit:
          type: string
          enum: [WORKSPACE_MONTH, GLOBAL_MONTH]
          description: >
            Present only when code is USAGE_LIMIT_REACHED. WORKSPACE_MONTH
            is this workspace's own monthly allowance, which starts again
            next calendar month; GLOBAL_MONTH is the allowance every
            workspace shares, which only whoever runs the service can raise.
        reason:
          type: string
          description: >
            Present only for some codes: with CONNECTOR_RESOURCE_UNAVAILABLE,
            GONE or CANCELLED for an event, and GONE, ACCESS_LOST, TRASHED,
            DOWNLOAD_RESTRICTED or CHANGED_DURING_COPY for a Drive file; with
            CONNECTOR_RESOURCE_UNSUPPORTED, TYPE or NOT_UTF8 for a Drive file
            (a calendar entry has none); with CONNECTION_RECONNECT_REQUIRED,
            TOKEN_REJECTED, TOKEN_UNREADABLE or PERMISSION_MISSING; with
            CONNECTOR_RESOURCE_REFUSED, the copy's rejection code from the
            same checks an upload meets; with ACTION_NOT_PROPOSABLE, NO_EXPORT,
            EXPORT_STALE, FORMAT_NOT_EXPORTED, FILE_TOO_LARGE,
            HIDDEN_CHARACTERS, TIME_SKIPPED or INVALID; with
            UNSUPPORTED_MEDIA_TYPE, why the file was refused: NOT_A_DOCUMENT,
            SPREADSHEET, PRESENTATION, PASSWORD_PROTECTED, RIGHTS_PROTECTED,
            REMOTE_CONTENT (it asks whoever opens it to fetch something from a
            network) or DAMAGED; with FILLABLE_FORM_FAILED, CANNOT_OPEN, DAMAGED
            or TIMED_OUT.
        access:
          type: string
          enum: [DRIVE_FILES, CALENDAR_EVENTS, DRIVE_SAVING, CALENDAR_EVENT_CREATION]
          description: >
            Present only when code is CONNECTION_RECONNECT_REQUIRED, naming the connection that must be made
            again, or CONNECTION_ACCESS_NOT_OFFERED, naming the connection nothing here would use.
        actionType:
          type: string
          enum: [DRIVE_SAVE_FILE, DRIVE_SAVE_AS_GOOGLE_DOC, CALENDAR_CREATE_EVENT, GOOGLE_DOC_APPEND]
          description: Present only when code is ACTION_NOT_OFFERED, naming the kind of change not made here.
        problems:
          type: array
          description: >
            Present only when code is RULE_VALIDATION_FAILED: one entry per
            way the proposed rule's payload does not hold against the
            draft's own field definitions.
          items:
            type: object
            properties:
              reason:
                type: string
              detail:
                type: string
...
    ApproveExportRequest:
      type: object
      required: [validationManifestId, format]
      properties:
        validationManifestId:
          type: integer
          format: int64
        format:
          type: string
          enum: [DOCX, PDF, BOTH]
    ExportApprovalResponse:
      type: object
      required: [id, documentId, revisionId, templateVersionId, validationManifestId, format, approvedAt]
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        revisionId:
          type: integer
          format: int64
        templateVersionId:
          type: integer
          format: int64
        validationManifestId:
          type: integer
          format: int64
        format:
          type: string
          enum: [DOCX, PDF, BOTH]
        approvedAt:
          type: string
          format: date-time
    ExportReceiptResponse:
      type: object
      required:
        [id, documentId, revisionId, templateVersionId, exportApprovalId, validationManifestId, docxArtifactId, docxSha256,
         format, isCompletePair, exportedAt]
      properties:
        id:
          type: integer
          format: int64
        documentId:
          type: integer
          format: int64
        revisionId:
          type: integer
          format: int64
        templateVersionId:
          type: integer
          format: int64
        exportApprovalId:
          type: integer
          format: int64
        validationManifestId:
          type: integer
          format: int64
        docxArtifactId:
          type: integer
          format: int64
          nullable: true
          description: Null for a PDF template's export, which is a PDF and nothing else.
        docxSha256:
          type: string
          nullable: true
        pdfArtifactId:
          type: integer
          format: int64
          nullable: true
        pdfSha256:
          type: string
          nullable: true
        format:
          type: string
          enum: [DOCX, PDF, BOTH]
        isCompletePair:
          type: boolean
        exportedAt:
          type: string
          format: date-time
...